refactor: change name and import badgeProfile
