+++ src/components/UserCard/index.tsx
@@ -1,33 +1,25 @@
-import { FiUser } from 'react-icons/fi';
+import { ProfileBadge } from '../ProfileBadge';
 
 import * as S from './styles';
 
-export interface UserModalProps {
+export interface UserCardProps {
   imageUrl: string;
   name: string;
   description: string;
   handleGotoProfile: () => void;
 }
 
-// TODO - RENAME TO USER CARD
-function UserModal({
+function UserCard({
   imageUrl,
   name,
   description,
   handleGotoProfile
-}: UserModalProps) {
+}: UserCardProps) {
   return (
     <S.Container onClick={handleGotoProfile} aria-label="Goto profile">
       <S.Content>
         <S.WrapperContent>
-          {/* @TODO - REFACTOR THIS, FOR NEW COMPONENT RULE */}
-          {imageUrl ? (
-            <S.ImageProfile src={imageUrl} alt="User image profile" />
-          ) : (
-            <S.Circle>
-              <FiUser size={32} color="fff" />
-            </S.Circle>
-          )}
+          <ProfileBadge imageUrl={imageUrl} imageAlt="User image profile" />
           <S.WrapperText>
             <S.Title>{name}</S.Title>
             <S.Description>{description}</S.Description>
@@ -38,4 +30,4 @@ function UserModal({
   );
 }
 
-export { UserModal };
+export { UserCard };
