+++ src/components/UserCard/styles.ts
